Move methods from Protocol to ServerProtocol
